Update script.py
The script saves the latest seen date of a post. The lastest seen date will be updated only if the post Is not changed in any way. If the seller of a product updates the price (or any other values of the post), the post's last seen date will not be updated. Instead, the newly updated post will get a new row In the .csv file.